src/backend/requirements.txt
sanic
sanic-ext
aiohttp
sqlalchemy[asyncio]
asyncpg
python-dotenv
